# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Industry Monitor API (勐腊数据采集)",
    "description": "勐腊行业数据监控与采集系统 API。支持行业数据查询、定时采集、缓存管理、告警监控等功能。",
    "version": "1.0.0",
    "contact": {
      "name": "Admin"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "本地开发环境"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer Token 认证。通过 /api/auth/login 获取 token。"
      }
    },
    "schemas": {
...
      "MengLaStatusRequest": {
        "type": "object",
        "required": ["granularity", "startDate", "endDate"],
        "properties": {
          "catId": { "type": "string", "description": "类目 ID" },
          "granularity": {
            "type": "string",
            "enum": ["day", "month", "quarter", "year"],
            "description": "粒度"
          },
          "startDate": { "type": "string", "format": "date", "description": "开始日期 (yyyy-MM-dd)" },
          "endDate": { "type": "string", "format": "date", "description": "结束日期 (yyyy-MM-dd)" },
          "actions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": ["high", "hot", "chance", "industryViewV2", "industryTrendRange"]
            },
            "description": "查询的动作列表"
          }
        }
      },
      "MengLaStatusResponse": {
        "type": "object",
        "properties": {
          "catId": { "type": "string" },
          "granularity": { "type": "string" },
          "startDate": { "type": "string" },
          "endDate": { "type": "string" },
          "status": {
            "type": "object",
            "description": "按 action → period_key → boolean 的嵌套状态",
            "additionalProperties": {
              "type": "object",
              "additionalProperties": { "type": "boolean" }
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": { "type": "string", "description": "错误详情" }
        }
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/admin/mengla/status": {
      "post": {
        "tags": ["管理运维"],
        "summary": "查询数据可用性",
        "description": "检查指定类目 + 日期范围 + 粒度的数据可用性状态",
        "operationId": "menglaStatus",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MengLaStatusRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MengLaStatusResponse" }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/admin/mengla/enqueue-full-crawl": {
      "post": {
        "tags": ["管理运维"],
        "summary": "提交全量爬取任务",
        "description": "创建基于队列的全量爬取任务",
        "operationId": "enqueueFullCrawl",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/EnqueueFullCrawlRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "crawl job enqueued" },
                    "jobId": { "type": "string" },
                    "startDate": { "type": "string" },
                    "endDate": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/admin/backfill": {
      "post": {
        "tags": ["管理运维"],
        "summary": "触发历史数据回填",
        "description": "触发历史数据回填任务（后台运行）",
        "operationId": "backfill",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BackfillRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "backfill started" },
                    "start_date": { "type": "string" },
                    "end_date": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/alerts": {
      "get": {
        "tags": ["管理运维 - 告警"],
        "summary": "获取活跃告警",
        "description": "获取当前活跃的告警列表",
        "operationId": "getAlerts",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "active": { "type": "array", "items": {}, "description": "活跃告警列表" },
                    "rules": { "type": "object", "description": "告警规则" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/admin/alerts/history": {
      "get": {
        "tags": ["管理运维 - 告警"],
        "summary": "获取告警历史",
        "description": "获取告警历史记录",
        "operationId": "getAlertsHistory",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "description": "返回记录数量限制",
            "schema": { "type": "integer", "default": 100 }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "object" },
                  "description": "告警历史记录"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/admin/alerts/check": {
      "post": {
        "tags": ["管理运维 - 告警"],
        "summary": "手动触发告警检查",
        "description": "手动触发一次告警检查",
        "operationId": "checkAlerts",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "告警检查结果"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/admin/alerts/silence": {
      "post": {
        "tags": ["管理运维 - 告警"],
        "summary": "静默告警规则",
        "description": "静默指定的告警规则",
        "operationId": "silenceAlert",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SilenceRuleRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Rule error_rate silenced for 60 minutes" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "规则未找到",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/admin/cache/stats": {
      "get": {
        "tags": ["管理运维 - 缓存"],
        "summary": "获取缓存统计",
        "description": "获取缓存统计信息",
        "operationId": "getCacheStats",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "缓存统计信息"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/admin/cache/warmup": {
      "post": {
        "tags": ["管理运维 - 缓存"],
        "summary": "缓存预热",
        "description": "触发缓存预热任务（后台运行）",
        "operationId": "cacheWarmup",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CacheWarmupRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Cache warmup started" },
                    "limit": { "type": "integer" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/admin/cache/clear-l1": {
      "post": {
        "tags": ["管理运维 - 缓存"],
        "summary": "清除 L1 缓存",
        "description": "清除本地 L1 缓存",
        "operationId": "clearL1Cache",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "L1 cache cleared" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/admin/circuit-breakers": {
      "get": {
        "tags": ["管理运维"],
        "summary": "获取熔断器状态",
        "description": "获取所有熔断器的当前状态",
        "operationId": "getCircuitBreakers",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "熔断器状态统计"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/admin/circuit-breakers/reset": {
      "post": {
        "tags": ["管理运维"],
        "summary": "重置所有熔断器",
        "description": "重置所有熔断器状态",
        "operationId": "resetCircuitBreakers",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "All circuit breakers reset" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/admin/system/status": {
      "get": {
        "tags": ["管理运维"],
        "summary": "获取系统综合状态",
        "description": "获取系统综合状态，包括指标、缓存、告警、熔断器和调度器信息",
        "operationId": "getSystemStatus",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "metrics": { "type": "object", "description": "指标信息" },
                    "cache": { "type": "object", "description": "缓存信息" },
                    "alerts": {
                      "type": "object",
                      "properties": {
                        "active_count": { "type": "integer", "description": "活跃告警数" },
                        "rules": { "type": "object", "description": "告警规则" }
                      }
                    },
                    "circuit_breakers": { "type": "object", "description": "熔断器状态" },
                    "scheduler": {
                      "type": "object",
                      "properties": {
                        "running": { "type": "boolean", "description": "是否运行中" },
                        "jobs": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": { "type": "string" },
                              "name": { "type": "string" },
                              "next_run": { "type": "string" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/admin/scheduler/status": {
      "get": {
        "tags": ["管理运维 - 调度器"],
        "summary": "获取调度器状态",
        "description": "获取调度器运行状态详情",
        "operationId": "getSchedulerStatus",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "running": { "type": "boolean", "description": "是否运行中" },
                    "state": {
                      "type": "string",
                      "enum": ["stopped", "running", "paused"],
                      "description": "运行状态"
                    },
                    "total_jobs": { "type": "integer", "description": "任务总数" },
                    "active_jobs": { "type": "array", "items": {}, "description": "活跃任务" },
                    "paused_jobs": { "type": "array", "items": {}, "description": "暂停任务" },
                    "background_tasks": { "type": "integer", "description": "后台任务数" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}